+++ CHANGELOG.md
@@ -121,3 +121,4 @@
     It requires a new request and parsing. Use it only if you really need it and movie.genres is not enough.
 - Update genre retrieval and add interests fetching functionality
 - Add info_series creators list to MovieDetail for series creators
+- FIX unhandled exception for series without episodes (ex. tt35373097)