+++ CHANGELOG.md
@@ -115,3 +115,10 @@
 
 ## v0.5.7
 - **Bug Fixes:** Fixed bugs in get_filmography
+
+## v0.5.8
+- Adding new service `get_all_interests` Fetch all interests for a title using the provided IMDb ID.
+    It requires a new request and parsing. Use it only if you really need it and movie.genres is not enough.
+- Update genre retrieval and add interests fetching functionality
+- Add info_series creators list to MovieDetail for series creators
+- FIX unhandled exception for series without episodes (ex. tt35373097)